% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtBias_Options_functions.R
\name{biascoeff.options}
\alias{biascoeff.options}
\title{Options Controlling the bias coefficients parameters.}
\usage{
biascoeff.options(...)
}
\arguments{
\item{...}{using one or more arguments of the form name = value.
Existing values can be retrieved by supplying the names (as character strings) of the components as unnamed arguments.}
}
\description{
Functions to handle settings used by the bias coefficients computation.
}
\details{
Available options are
\itemize{ 
  \item{\code{mulBiasFunction}: }{character, the function to be used to compute the multiplicative bias coefficients. Options are "mean" (default), "median"}
  \item{\code{qmecdfBoxInterp}: }{character, the interpolation method to use in case of \strong{Quantile Mapping with Empirical Distribution}. Options are "idw" (default), "means"}
  \item{\code{qmecdfBoxMaxdist}: }{numeric, maximum distance to use if \code{qmecdfBoxInterp} is \code{TRUE}}
  \item{\code{qmdistRainyDayThres}: }{numeric, rainy day threshold to use when computing the occurrence of rainy days in case of \strong{Quantile Mapping with Fitted Distribution}}
  \item{\code{blockType}: }{character, the method to use to create the block estimation when the argument \code{use.block} from \code{interp.method} is \code{TRUE}.
         Available options are \code{"gaussian"} and \code{"userdefined"}
         \itemize{
            \item{\code{"gaussian"}: }{using Gaussian quadrature method}
            \item{\code{"userdefined"}: }{user defined block}
       }
   }
   \item{\code{blockSize}: }{vector, the size of the block.
         \itemize{
            \item{\code{"gaussian"}: }{vector of length 2 in the form \code{c(width_x, width_y)}}
            \item{\code{"userdefined"}: }{vector of length 4 in the form \code{c(width_x, by_x, width_y, by_y)}}
       }
   }
  \item{\code{addCoarseGrid}: }{logical, add a coarse grid to interpolate the multiplicative bias for IDW and Ordinary Kriging method.
         The coarse grid will be created by resampling the grid of the input data and providing the resolution with \code{resCoarseGrid}.
         The bias values at the coarse grid will be set to 1}
  \item{\code{resCoarseGrid}: }{numeric, the resolution of the coarse grid in degree decimal.}
}
}
