% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtBlanking_Options_functions.R
\name{blanking.options}
\alias{blanking.options}
\title{Options Controlling the blanking buffer outside the polygons boundaries.}
\usage{
blanking.options(...)
}
\arguments{
\item{...}{using one or more arguments of the form name = value.
Existing values can be retrieved by supplying the names (as character strings) of the components as unnamed arguments.}
}
\description{
Functions to handle the width of buffer outside the polygons boundaries.
}
\details{
Available options are
\itemize{ 
   \item{\code{bufferOption}: }{character, options are \code{"default"} or \code{"user"}.
        \itemize{
             \item{\code{"default"}: }{a buffer of 4 times of the spatial resolution of the gridded data is used}
             \item{\code{"user"}: }{the width of the buffer will be defined by the user in \code{bufferWidth}}
           }
        }
   \item{\code{bufferWidth}: }{numeric, the width of the buffer outside the polygons in decimal degree}
}
}
