% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtCompute_Seasonal_Cmd.R
\name{cdtComputeSeasonal_cdtstn}
\alias{cdtComputeSeasonal_cdtstn}
\title{Computing seasonal data for CDT stations data.}
\usage{
cdtComputeSeasonal_cdtstn(
  season.def = list(start.month = 1, season.length = 3),
  period = list(start.year = 1981, end.year = 2010),
  station.data = list(time.step = "dekadal", file = "", sep = ",", missing = "-99"),
  season.min.frac = 0.95,
  aggregation.fun = "sum",
  count.opr = ">=",
  count.thres = 1,
  ...
)
}
\arguments{
\item{season.def}{a named list, season definition.}

\item{period}{a named list, the period of data to compute.}

\item{station.data}{a named list, stations data info.}

\item{season.min.frac}{used to aggregate the seasonal data, minimum fraction of non-missing values for each season.}

\item{aggregation.fun}{character, aggregation function to use. Options are: "sum", "mean", "median", "max", "min", "sd", "var", "count", "user".}

\item{count.opr}{character, the comparison operator to use when \code{aggregation.fun} is \code{"count"}.}

\item{count.thres}{the threshold to use when \code{aggregation.fun} is \code{"count"}.}

\item{...}{when \code{aggregation.fun} is \code{"user"}, the user function and additional arguments to be passed to the user's function.}
}
\value{
A named list
}
\description{
Compute seasonal data for CDT stations data.
}
