% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtCompute_Seasonal_Cmd.R
\name{cdtComputeSeasonal_netcdf}
\alias{cdtComputeSeasonal_netcdf}
\title{Computing seasonal data for netCDF dataset.}
\usage{
cdtComputeSeasonal_netcdf(
  season.def = list(start.month = 1, season.length = 3),
  period = list(start.year = 1981, end.year = 2010),
  netcdf.data = list(time.step = "dekadal", dir = "", format = "tmax_mrg_\%s\%s\%s.nc",
    varid = "temp", ilon = 1, ilat = 2),
  season.min.frac = 0.95,
  aggregation.fun = "mean",
  count.opr = ">=",
  count.thres = 1,
  ...
)
}
\arguments{
\item{season.def}{named list, season definition.
\itemize{
  \item{\code{start.month}: }{integer, start month of the season}
  \item{\code{season.length}: }{integer, length of the season}
}}

\item{period}{named list, the period of data to compute.
\itemize{
  \item{\code{start.year}: }{integer, start year of the period}
  \item{\code{end.year}: }{integer, the end year}
}}

\item{netcdf.data}{named list, netCDF data info.
\itemize{
\item{\code{time.step}: }{character, the time step of the netCDF data. Available options: \code{"daily"}, \code{"pentad"}, \code{"dekadal"}, \code{"monthly"}}
\item{\code{dir}: }{character, full path to the directory containing the netCDF files.}
\item{\code{format}: }{character, format of the netCDF file names}
\item{\code{varid}: }{character, name of the variable to read from the netCDF data}
\item{\code{ilon}: }{integer, order for the longitude dimension of the variable. 
Example: if the variable "precip" has the dimension order [Lat, Lon] then \code{ilon} must be 2}
\item{\code{ilat}: }{integer, order for the latitude dimension of the variable.}
}}

\item{season.min.frac}{numeric, used to aggregate the seasonal data, minimum fraction of non-missing values for each season.}

\item{aggregation.fun}{character, aggregation function to use. Options are: "sum", "mean", "median", "max", "min", "sd", "var", "count", "user".}

\item{count.opr}{character, the comparison operator to use when \code{aggregation.fun} is \code{"count"}.}

\item{count.thres}{the threshold to use when \code{aggregation.fun} is \code{"count"}.}

\item{...}{when \code{aggregation.fun} is \code{"user"}, the user function and additional arguments to be passed to the user's function.}
}
\value{
A named list
}
\description{
Compute seasonal data for netCDF dataset.
}
