% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtCrossValidation_ClimData_Cmd.R
\name{cdtCrossValidationClimDataCMD}
\alias{cdtCrossValidationClimDataCMD}
\title{Cross-validation, merging stations observation and gridded climate data.}
\usage{
cdtCrossValidationClimDataCMD(
  variable = "temp",
  time.step = "dekadal",
  minhour = 1,
  dates = list(from = "range", pars = list(start = "2018011", end = "2018123")),
  station.data = list(file = "", sep = ",", na.strings = "-99"),
  netcdf.data = list(dir = "", format = "tmax_adj_\%s\%s\%s.nc", varid = "temp", ilon =
    1, ilat = 2),
  merge.method = list(method = "SBA", nrun = 3, pass = c(1, 0.75, 0.5)),
  interp.method = list(method = "idw", nmin = 8, nmax = 16, maxdist = 2.5, use.block =
    TRUE, vargrd = FALSE, vgm.model = c("Sph", "Exp", "Gau", "Pen")),
  crossv.station = list(from = "all", pars = NULL),
  auxvar = list(dem = FALSE, slope = FALSE, aspect = FALSE, lon = FALSE, lat = FALSE),
  dem.data = list(file = "", varid = "dem", ilon = 1, ilat = 2),
  RnoR = list(use = FALSE, wet = 1, smooth = FALSE),
  output.dir = "",
  GUI = FALSE
)
}
\arguments{
\item{variable}{character, name of the climate variable to merge. Available options: \code{"rain"}, \code{"temp"}, \code{"rh"}, \code{"pres"}, \code{"prmsl"}, \code{"rad"}, \code{"wspd"}, \code{"ugrd"}, \code{"vgrd"}.
\itemize{
\item \code{"rain"}: rainfall data
\item \code{"temp"}: temperature data
\item \code{"rh"}: relative humidity data
\item \code{"pres"}: surface pressure data
\item \code{"prmsl"}: pressure at mean sea level
\item \code{"rad"}: radiation data
\item \code{"wspd"}: wind speed data
\item \code{"ugrd"}: zonal wind components, E-W speed
\item \code{"vgrd"}: meridional wind components, N-S speed
}}

\item{time.step}{character, the time step of the data. Available options: \code{"minute"}, \code{"hourly"}, \code{"daily"}, \code{"pentad"}, \code{"dekadal"}, \code{"monthly"}.}

\item{minhour}{integer, the step for \code{"minute"} and \code{"hourly"}.
\itemize{
\item \code{"minute"}: can have the value 1, 5, 10, 15, 20 or 30
\item \code{"hourly"}: can take the value 1, 3, 6 or 12
}}

\item{dates}{named list, providing the dates to merge.
The list includes an element \code{from} with available options \code{"range"}, \code{"file"} or \code{"dates"}, and 
 an element \code{pars} which is a named list specifying the parameters related to \code{from}:
\itemize{
\item \strong{"range"}: \code{pars} specifies the start and end dates to merge. \cr
Example: \code{pars = list(start = "2018011", end = "2018123")}
\item \strong{"file"}: \code{pars} specifies the full path to the file containing the dates to merge.
 Example: \code{pars = list(file = "/home/data/files/dates.txt")}\cr
The contents of the file are as follows:\cr
## cat /home/data/files/dates.txt \cr
2020011\cr
2020012\cr
2020013\cr
......
\item \strong{"dates"}: \code{pars} specifies a vector containing the dates to merge. \cr
Example: \code{pars = list(dates = c("2020011", "2020012", 2020091, 2020113))}
}}

\item{station.data}{named list, providing the station data to be used in CDT format.
\itemize{
\item \code{file}: character, full path to the file containing the stations data
\item \code{sep}: character, column separator of the data
\item \code{na.strings}: character, missing values flag
}}

\item{netcdf.data}{named list, providing the input netCDF dataset to be used.
\itemize{
\item \code{dir}: character, full path to the directory containing the netCDF files.
\item \code{format}: character, format of the netCDF file names
\item \code{varid}: character, name of the variable to read from the netCDF data
\item \code{ilon}: integer, order for the longitude dimension of the variable. 
Example: if the variable "precip" has the dimension order [Lat, Lon] then \code{ilon} must be 2
\item \code{ilat}: integer, order for the latitude dimension of the variable.
}}

\item{merge.method}{named list, indicating the merging method.
\itemize{
\item \code{"method"}: character, the merging method. Valid options: \code{"CSc"}, \code{"BSc"}, \code{"SBA"} or \code{"RK"}.
 \itemize{
  \item \strong{"CSc"}: Cressman Scheme
  \item \strong{"BSc"}: Barnes Scheme
  \item \strong{"SBA"}: Simple Bias Adjustment
  \item \strong{"RK"}: Regression Kriging
 }
\item \code{"nrun"}: integer, number of the nested run to be performed
\item \code{"pass"}: numeric vector giving the fraction of \code{nmin}, \code{nmax} and \code{maxdist} to be used for each pass.
}}

\item{interp.method}{named list, indicating the interpolation method and parameters to be used for \code{"SBA"} and \code{"RK"}. 
\itemize{
\item \code{method}: character, the interpolation method to be used.
Valid options: \code{"idw"}, \code{"shepard"}, \code{"sphere"} or \code{"okr"}.
\itemize{
 \item \strong{"idw"}: Inverse distance weighted
 \item \strong{"shepard"}: Modified Shepard interpolation
 \item \strong{"sphere"}: Spheremap interpolation method
 \item \strong{"okr"}: Ordinary kriging
}
\item \code{nmin}: integer, minimum number of stations to be used to interpolate a grid point
\item \code{nmax}: integer, maximum number of stations to be used to interpolate a grid point
\item \code{maxdist}: numeric, maximum radius of influence in decimal degree
\item \code{use.block}: logical, use block mean values to interpolate a grid point
\item \code{vargrd}: logical, use a variable radius of influence
\item \code{vgm.model}: character vector of variogram model to be used if \code{method} is \code{"okr"}. Default is \code{c("Exp", "Gau", "Sph", "Pen")}
}}

\item{crossv.station}{named list, selecting the stations to use for the cross-validation.
The list includes an element \code{from} with available options \code{"all"}, \code{"file"} or \code{"cdt"}, and 
an element \code{pars} which is a named list specifying the parameters related to \code{from}:
\itemize{
\item \strong{"all"}: all the stations from \code{station.data} will be used for cross-validation, \code{pars} can be omitted
\item \strong{"file"}: the list of stations to be used for the cross-validation comes from a file.
 \itemize{
   \item \code{type}: character, the type of the data, valid options are:\cr
      \strong{"cdtstation"}: the stations come from a CDT stations data, \cr
      \strong{"cdtcoords"}: the stations come from a CDT coordinates file
   \item \code{file}: character, full path to the file containing the stations
   \item \code{sep}: character, column separator of the data
   \item \code{na.strings}: character, missing values flag
   \item \code{header}: logical, in case of \strong{"cdtcoords"}, set \code{TRUE} if the data has a header
 }
\item \strong{"cdt"}: the list of stations to be used for the cross-validation will be selected from \code{station.data} 
  by providing the percent of minimum available data for each station.\cr
  Example: \code{pars = list(min.perc = 40)}
}}

\item{auxvar}{named list, specifying the auxiliary variables to use when the merging method is \code{"RK"}.
\itemize{
\item \code{dem}: logical, include elevation data as auxiliary variable
\item \code{slope}: logical, include slope as auxiliary variable
\item \code{aspect}: logical, include aspect as auxiliary variable
\item \code{lon}: logical, include longitude as auxiliary variable
\item \code{lat}: logical, include latitude as auxiliary variable
}}

\item{dem.data}{named list, providing the Digital Elevation Model (in netCDF format) when using regression kriging method with elevation related data as auxiliary variable.
\itemize{
\item \code{file}: character, full path to the netCDF file containing the elevation data.
\item \code{varid}: character, name of the variable to read from the netCDF data
\item \code{ilon}: integer, order for the longitude dimension of the variable.
\item \code{ilat}: integer, order for the latitude dimension of the variable.
}}

\item{RnoR}{named list, specifying the rain-no-rain mask parameters in case of \code{var.clim} = \code{"rain"}.
\itemize{
\item \code{use}: logical, apply rain-no-rain mask
\item \code{wet}: numeric, threshold to be use to define the wet/dry event
\item \code{smooth}: logical, smooth the rain-no-rain mask after interpolation
}}

\item{output.dir}{character, full path to the directory to save the output.}

\item{GUI}{logical, indicating whether or not the output message should be displayed on CDT GUI. If \code{TRUE}, CDT GUI must be open.}
}
\description{
Function to perform a Leave-One-Out Cross-Validation for climate data merging.
}
