% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtCreateDataset_NCmultiDim.R
\name{cdtDataset_readData_multiDim}
\alias{cdtDataset_readData_multiDim}
\title{Create CDT dataset format.}
\usage{
cdtDataset_readData_multiDim(
  cdtdata_info,
  netcdf_data,
  ncvar_info,
  extra_ncdim = NULL,
  conversion_fun = NULL,
  chunk = list(size = 100, fac = 2)
)
}
\arguments{
\item{cdtdata_info}{named list, providing the information about the CDT dataset.
\itemize{
\item{\code{dir}: }{character, full path to the directory to save the created CDT dataset.}
\item{\code{name}: }{character, name of the dataset to use.}
}}

\item{netcdf_data}{named list, providing the input netCDF data to be converted to a CDT dataset format.
\itemize{
\item{\code{timestep}: }{character, the time step of the netCDF data.}
\item{\code{dir}: }{character, full path to the directory containing the netCDF files.}
\item{\code{format}: }{character, format of the netCDF file names. Example: \code{'precip_\%S.nc'}}
}}

\item{ncvar_info}{named list, providing the name of the variable and longitude, latitude, time of the netCDF data.
\itemize{
\item{\code{varid}: }{character, the name of the variable to be converted.}
\item{\code{lon_dim}: }{character, the name of the longitude dimension.}
\item{\code{lat_dim}: }{character, the name of the latitude dimension.}
\item{\code{time_dim}: }{character, the name of the time dimension.}
}}

\item{extra_ncdim}{named list, providing the values of the dimensions other than the longitude,
latitude and time to be extracted if the netCDF data contains more than 3 dimensions.
Default \code{NULL}, the data does not have extra dimension.}

\item{conversion_fun}{a function to transform the data or to convert the units.}

\item{chunk}{named list containing the size of the chunk \code{size} and the number of chunk to be used for each computation \code{fac}.}
}
\value{
A directory named after the element \code{name} of \code{cdtdata_info}, containing the CDT dataset.
}
\description{
Create CDT dataset from a multidimensional netCDF data.
}
\examples{

\dontrun{

cdtdata_info <- list(dir = "~/CDTDATASET/MERRA2", name = "PRECIP")
netcdf_data <- list(timestep = 'daily', dir = "~/DATA/MERRA2/daily_precip", format = 'merra-2_precip_\%S.nc')
ncvar_info <- list(varid = 'precip', lon_dim = 'lon', lat_dim = 'lat', time_dim = 'time')
convert_units <- function(x){
   # kg m-2 s-1 to mm day-1
   x * 86400
}

CDT::cdtDataset_readData_multiDim(cdtdata_info, netcdf_data, ncvar_info, conversion_fun = convert_units)

}

}
