% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtInstall_Packages.R
\name{cdtInstall.Packages}
\alias{cdtInstall.Packages}
\title{Check and install CDT dependencies packages}
\usage{
cdtInstall.Packages(interactive = FALSE, config = NULL)
}
\arguments{
\item{interactive}{if \code{TRUE} the path to the packages directory (Windows)
or configure.args for \code{rgeos} and \code{rgdal} (MacOS and Linux)
will be entered from the console.}

\item{config}{A list containing the full path of the packages directory or
the path to set the configure.args for \code{rgeos} and \code{rgdal}.
Leave \code{NULL} if the \code{interactive} is \code{TRUE}.
(See example below.)}
}
\description{
Check CDT dependencies packages and install them if not installed yet
}
\examples{
library(CDT)

# Interactive mode
cdtInstall.Packages()

# Windows
cdtInstall.Packages(FALSE, config = list(
		local.dir = TRUE,
		pkg.path = "C:/Users/rijaf/Desktop/R3.5-packages")
 )

# MacOS and Linux
cdtInstall.Packages(FALSE, config = list(
		geos_config = "/Library/Frameworks/GEOS.framework/unix/bin/geos-config",
		gdal_config = "/Library/Frameworks/GDAL.framework/unix/bin/gdal-config",
		proj_include = "/Library/Frameworks/PROJ.framework/unix/include",
		proj_lib = "/Library/Frameworks/PROJ.framework/unix/lib")
 )

}
