% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtInterpolation_functions.R
\name{cdtInterp}
\alias{cdtInterp}
\title{CDT interpolation methods}
\usage{
cdtInterp(locations, values, newdata, nmin = 4, nmax = 10,
  method = "idw", spheric = FALSE, vgm = NULL, p = 2)
}
\arguments{
\item{locations}{A matrix containing the coordinates of the station data. The 1st column is the longitude and the 2nd coulmn the latitude.}

\item{values}{A vector containing the values of the station. The length of \code{values} must be equal to the number of row of \code{locations}.}

\item{newdata}{A matrix containing the coordinates of the grid to interpolate. The 1st column is the longitude and the 2nd coulmn the latitude.}

\item{nmin}{The minimum number of stations to be used to interpolate a grid point.}

\item{nmax}{The maximum number of stations to be used to interpolate a grid point.}

\item{method}{Method to calculate weights for the interpolation. Should be \strong{"idw"} (default), \strong{"shepard"}, \strong{"spheremap"} or \strong{"kriging"}.
\itemize{
 \item \strong{"idw"}: Inverse distance weighted
 \item \strong{"shepard"}: Modified Shepard interpolation
 \item \strong{"spheremap"}: Spheremap interpolation method
 \item \strong{"kriging"}: Ordinary kriging
}}

\item{spheric}{If \code{FALSE} (default), then a Cartesian distance will be computed. If set to \code{TRUE}, a spherical distance will be computed.}

\item{vgm}{A \code{gstat} variogram model. Example: \code{vgm(10.2, "Sph", 2.7, 1.1)}.}

\item{p}{The power to use in weight calculation.}
}
\value{
A matrix containing the interpolated data. The 1st and 2nd columns are the same as \code{newdata} and the 3rd column contains the interpolated values.
}
\description{
This function interpolates stations data on a rectangular grid in the x-y plane.
The radius of influence for the interpolation method varies for each grid point.
}
