% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtDownloadRFE_mswep.R
\name{connect_tidyverse_api}
\alias{connect_tidyverse_api}
\title{Connect Tidyverse API to Google Drive.}
\usage{
connect_tidyverse_api(email)
}
\arguments{
\item{email}{email for the Google account.}
}
\description{
Connect Tidyverse API to Google Drive.
}
