% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtCPTconversion_Procs.R
\name{convert2CPTFormat}
\alias{convert2CPTFormat}
\title{Conversion to CPT data format.}
\usage{
convert2CPTFormat(
  data.type = "cdtstation",
  station.data = list(file = "", sep = ",", na.strings = "-99"),
  netcdf.data = list(dir = "", format = "onset_\%Y\%M\%D.nc", varid = "onset", ilon =
    1, ilat = 2),
  cpt.varinfo = list(name = "onset", units = "days since", missval = "-9999"),
  output.file = "",
  GUI = FALSE
)
}
\arguments{
\item{data.type}{character, the type of the data. Available options: \code{"cdtstation"} or \code{"cdtnetcdf"}.}

\item{station.data}{named list, providing the station data to convert in CDT format.
\itemize{
\item{\code{file}: }{character, full path to the file containing the stations data}
\item{\code{sep}: }{character, column separator of the data}
\item{\code{na.strings}: }{character, missing values flag}
}}

\item{cpt.varinfo}{named list, information about the field of the output CPT file.
\itemize{
\item{\code{name}: }{character, the field name}
\item{\code{units}: }{character, the units}
\item{\code{missval}: }{character, missing values flag}
}}

\item{output.file}{character, full path to the file to save the CPT data format.}

\item{GUI}{logical, indicating whether or not the output message should be displayed on CDT GUI. If \code{TRUE}, CDT GUI must be open.}
}
\description{
Function to correct station data or netCDF files to CPT data format.
}
