% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Variables_functions.R
\name{daylight_hours}
\alias{daylight_hours}
\title{Compute the daylight hours.}
\usage{
daylight_hours(lat, dates)
}
\arguments{
\item{lat}{numeric, a vector of the latitude of the points where the daylight hours will be calculated.}

\item{dates}{character, a vector of dates in the form of \code{YYYYMMDD}.}
}
\value{
Returns a matrix, where the column represent the latitude and the row the date.
}
\description{
Compute the daylight hours for different latitudes.
}
\references{
Allen, R.G., Pereira, L.S., Raes, D., and Smith, M. 1998. Crop evapotranspiration: Guidelines for computing crop requirements. Irrigation and Drainage PaperNo. 56, FAO, Rome, Italy 300 p.
}
