% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Humidity_functions.R
\name{dewpoint_temperature}
\alias{dewpoint_temperature}
\title{Dewpoint temperature.}
\usage{
dewpoint_temperature(tm, rh)
}
\arguments{
\item{tm}{numeric, a vector or matrix of the air temperature in degree Celsius.}

\item{rh}{numeric, a vector or matrix of the relative humidity in percentage.}
}
\value{
Returns a vector or matrix, the units are in degree Celsius.
}
\description{
Compute the dewpoint temperature from the air temperature and relative humidity.
}
