% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtSrcDistance_functions.R
\name{distance.Matrix}
\alias{distance.Matrix}
\title{Distance matrix}
\usage{
distance.Matrix(x, y, spheric = FALSE)
}
\arguments{
\item{x}{a 2 column matrix or data.frame (first column is longitude, second is latitude); or a SpatialPoints*, SpatialPixels* or SpatialGrid* object.}

\item{y}{Same as \code{x}. If \code{x} is a Spatial object, \code{y} can take a different Spatial object other than \code{x}.}

\item{spheric}{If \code{FALSE} (default), then a Cartesian distance will be computed. If set to \code{TRUE}, a spherical distance (using a standard great circle method) will be computed.}
}
\value{
Matrix of distances. The row represent \code{y} and the column \code{x}.
}
\description{
Distance matrix between two sets of points
}
