% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtUpdate_ONI_function.R
\name{download.oni.cpc.ncep.noaa}
\alias{download.oni.cpc.ncep.noaa}
\title{Update Oceanic Niño Index (ONI)}
\usage{
download.oni.cpc.ncep.noaa(GUI = FALSE)
}
\arguments{
\item{GUI}{logical, if TRUE, display message on CDT (if CDT GUI is open), otherwise the message is displayed on R console.}
}
\description{
Updating CDT's ONI data from cpc.ncep.noaa.gov.
}
\references{
\url{https://origin.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ONI_v5.php}
}
