% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_ET0_FAO_functions.R
\name{ET0_Penman_Monteith_FAO_station}
\alias{ET0_Penman_Monteith_FAO_station}
\title{Compute daily reference evapotranspiration for CDT station data format.}
\usage{
ET0_Penman_Monteith_FAO_station(
  tmax.data = list(file = "", sep = ",", na.strings = "-99"),
  tmin.data = list(file = "", sep = ",", na.strings = "-99"),
  rh.from = "minmax",
  rhmax.data = list(file = "", sep = ",", na.strings = "-99"),
  rhmin.data = list(file = "", sep = ",", na.strings = "-99"),
  rhmean.data = list(file = "", sep = ",", na.strings = "-99"),
  wff10.data = list(file = "", sep = ",", na.strings = "-99"),
  pres.data = list(file = "", sep = ",", na.strings = "-99"),
  rad.data = list(file = "", sep = ",", na.strings = "-99"),
  elev.from = "netcdfDEM",
  elev.data = list(file = "", varid = "z", ilon = 1, ilat = 2),
  output = list(file = "", sep = ",", na.strings = "-99")
)
}
\arguments{
\item{tmax.data}{named list, providing the maximum temperature data in CDT station data format. Units: degree Celsius.
\itemize{
\item{\code{file}: }{character, full path to the file containing the stations data}
\item{\code{sep}: }{character, column separator of the data}
\item{\code{na.strings}: }{character, missing values flag}
}}

\item{tmin.data}{named list, providing the minimum temperature data in CDT station data format. Units: degree Celsius.
See \code{tmax.data} for the elements of the list.}

\item{rh.from}{character, source of relative humidity data. Valid options: \code{"minmax"} and \code{"mean"}.
\itemize{
\item{\code{"minmax"}: }{the humidity data are from the minimum and maximum relative humidity.}
\item{\code{"mean"}: }{the humidity data is from a mean relative humidity data.}
}}

\item{rhmax.data}{named list, if \code{rh.from} is equal to \code{"minmax"},
providing the maximum relative humidity data in CDT station data format. Units: percentage.
See \code{tmax.data} for the elements of the list.}

\item{rhmin.data}{named list, if \code{rh.from} is equal to \code{"minmax"},
providing the minimum relative humidity data in CDT station data format. Units: percentage.
See \code{tmax.data} for the elements of the list.}

\item{rhmean.data}{named list, if \code{rh.from} is equal to \code{"mean"},
providing the mean relative humidity data in CDT station data format. Units: percentage.
See \code{tmax.data} for the elements of the list.}

\item{wff10.data}{named list, providing the wind speed at 10 meters data in CDT station data format. Units: m/s.
See \code{tmax.data} for the elements of the list.}

\item{pres.data}{named list, providing the surface pressure data in CDT station data format. Units: hPa.
See \code{tmax.data} for the elements of the list.}

\item{rad.data}{named list, providing the solar radiation flux data in CDT station data format. Units: W/m^2.
See \code{tmax.data} for the elements of the list.}

\item{elev.from}{character, source of the elevation data. Valid options: \code{"inputPresData"}, \code{"cdtCrdFile"} and \code{"netcdfDEM"}.
\itemize{
\item{\code{"inputTmaxData"}: }{the elevation data is from the maximum temperature data.}
\item{\code{"cdtCrdFile"}: }{the elevation data is from a CDT coordinates file. The ID must be the same as the maximum temperature data.}
\item{\code{"netcdfDEM"}: }{the elevation data is from the pressure data.}
}
Default is \code{"netcdfDEM"}.}

\item{elev.data}{named list, if \code{elev.from} is equal to \code{"cdtCrdFile"},
set the elements of the list providing the CDT coordinate file containing the elevation data. Units: meters.
\itemize{
\item{\code{file}: }{character, full path to the CDT coordinate file}
\item{\code{sep}: }{character, column separator of the data}
\item{\code{na.strings}: }{character, missing values flag}
}
If \code{elev.from} is equal to \code{"netcdfDEM"} set the elements of the list providing the Digital Elevation Model in netCDF format. Units: meters.
\itemize{
\item{\code{file}: }{character, full path to the netCDF file containing the elevation data.}
\item{\code{varid}: }{character, name of the variable to read from the netCDF data.}
\item{\code{ilon}: }{integer, order for the longitude dimension of the variable.}
\item{\code{ilat}: }{integer, order for the latitude dimension of the variable.}
}}

\item{output}{named list, the elements of the list are
\itemize{
\item{\code{file}: }{character, full path to the file to save the calculated reference evapotranspiration}
\item{\code{sep}: }{character, column separator of the data}
\item{\code{na.strings}: }{character, missing values flag}
}}
}
\description{
Function to compute daily reference evapotranspiration using the FAO Penman-Monteith equation.
}
\references{
Allen, R.G., Pereira, L.S., Raes, D., and Smith, M. 1998. Crop evapotranspiration: Guidelines for computing crop requirements. Irrigation and Drainage PaperNo. 56, FAO, Rome, Italy 300 p.
\url{https://www.fao.org/3/X0490E/x0490e00.htm}
}
