% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtExtractData_Geom.R
\name{extractGeomPoints}
\alias{extractGeomPoints}
\title{Geom points extraction}
\usage{
extractGeomPoints(gridObj, points, padxy)
}
\arguments{
\item{gridObj}{sp SpatialPixels}

\item{points}{data.frame(name, longitude, latitude)}

\item{padxy}{c(padx, pady) in decimal degree}
}
\value{
A named list
}
\description{
Geom points extraction.
}
