% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtExtractData_Geom.R
\name{extractGeomPolys}
\alias{extractGeomPolys}
\title{Geom polygons extraction}
\usage{
extractGeomPolys(gridObj, shpf, attr_name, attr_values)
}
\arguments{
\item{gridObj}{sp SpatialPixels}

\item{shpf}{sp SpatialPolygonsDataFrame (from shapefile)}

\item{attr_name}{attribute name, 'GID_1'}

\item{attr_values}{vector of the polygons to extract, c('RWA.1_1', 'RWA.5_1')}
}
\value{
A named list
}
\description{
Geom polygons extraction.
}
