% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Variables_functions.R
\name{extraterrestrial_radiation_annual}
\alias{extraterrestrial_radiation_annual}
\title{Compute the extraterrestrial radiation.}
\usage{
extraterrestrial_radiation_annual(lat, tstep = "daily")
}
\arguments{
\item{lat}{numeric, a vector of the latitude of the points where the extraterrestrial radiation will be calculated.}

\item{tstep}{character, the time step of the computed extraterrestrial radiation. Available options: \code{"daily"}, \code{"pentad"}, \code{"dekadal"}, \code{"monthly"}.}
}
\value{
Returns a matrix, where the column represent the latitude and the row the time of the year. The number of row for \code{"daily"} is 365, 72 for \code{"pentad"}, 36 for \code{"dekadal"} and 12 \code{"monthly"}.
}
\description{
Compute the extraterrestrial radiation throughout the year for different latitudes.
}
\references{
Allen, R.G., Pereira, L.S., Raes, D., and Smith, M. 1998. Crop evapotranspiration: Guidelines for computing crop requirements. Irrigation and Drainage PaperNo. 56, FAO, Rome, Italy 300 p.
}
