% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Variables_functions.R
\name{extraterrestrial_radiation_daily}
\alias{extraterrestrial_radiation_daily}
\title{Compute the extraterrestrial radiation for daily data.}
\usage{
extraterrestrial_radiation_daily(lat, dates)
}
\arguments{
\item{lat}{numeric, a vector of the latitude of the points where the extraterrestrial radiation will be calculated.}

\item{dates}{character, a vector of dates in the form of \code{YYYYMMDD}.}
}
\value{
Returns a matrix, where the column represent the latitude and the row the date.
}
\description{
Compute the extraterrestrial radiation of daily data for different latitudes.
}
\references{
Allen, R.G., Pereira, L.S., Raes, D., and Smith, M. 1998. Crop evapotranspiration: Guidelines for computing crop requirements. Irrigation and Drainage PaperNo. 56, FAO, Rome, Italy 300 p.
}
