% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtInstall_wgrib.R
\name{install_wgrib}
\alias{install_wgrib}
\title{Install wgrib.}
\usage{
install_wgrib()
}
\description{
Function to download and install wgrib into CDT local folder.
}
\references{
\url{https://www.cpc.ncep.noaa.gov/products/wesley/wgrib.html}
}
