% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtInstall_wgrib.R
\name{install_wgrib2}
\alias{install_wgrib2}
\title{Install wgrib2.}
\usage{
install_wgrib2(make = "make", cc = "gcc", fc = "gfortran")
}
\arguments{
\item{make}{the make command or full path to make command.}

\item{cc}{the \code{C} compiler to be used or the full path to the \code{C} compiler.}

\item{fc}{the \code{Fortran} compiler to be used or the full path to the \code{Fortran} compiler.}
}
\description{
Function to download and build wgrib2 into CDT local folder.
}
\examples{

\dontrun{
## Install wgrib2 on Windows
install_wgrib2()

## Install wgrib2 on MacOS
install_wgrib2(make = '/usr/local/bin/gmake', cc = '/usr/local/bin/gcc-13', fc = '/usr/local/bin/gfortran')

## Install wgrib2 on Linux with the default make, CC and FC
install_wgrib2()
}

}
\references{
\url{https://www.cpc.ncep.noaa.gov/products/wesley/wgrib2/index.html}
}
