% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtDownloadReanalysis_jra55_apps.R
\name{jra55.format.ncdf}
\alias{jra55.format.ncdf}
\title{Splitting and formatting the JRA55 3 hourly downloaded data.}
\usage{
jra55.format.ncdf(dirNCDF, dirOUT)
}
\arguments{
\item{dirNCDF}{full path to the folder containing the downloaded netCDF files.
If the data are compressed wit "tar" or "zip", make sure to untar or unzip all files
and put the uncompressed netCDF files under this folder}

\item{dirOUT}{full path to the folder you want to save the formatted data.}
}
\description{
Function to split and format the downloaded JRA55 minimum and maximum temperature at 2 meter above ground.
}
