% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtDownloadReanalysis_jra55_apps.R
\name{jra55.send.request}
\alias{jra55.send.request}
\title{Submitting subset requests for JRA55 minimum and maximum temperature at 2 meter above ground.}
\usage{
jra55.send.request(
  variable = "tmax",
  email = NULL,
  password = NULL,
  bbox = list(minlon = 42, maxlon = 52, minlat = -26, maxlat = -11),
  start_time = list(year = 2020, month = 11, day = 1, hour = 0),
  end_time = list(year = 2020, month = 11, day = 30, hour = 21)
)
}
\arguments{
\item{variable}{the variable to be requested, available options are "tmax": maximum temperature and "tmin" minimum temperature.}

\item{email}{your email address used as login on https://rda.ucar.edu/.}

\item{password}{your password.}

\item{bbox}{a named list providing the bounding box of the area to extract. The names of components must minlon, maxlon, minlat and maxlat.}

\item{start_time}{a named list providing the start time to extract. The names of components must year, month, day and hour}

\item{end_time}{a named list providing the end time to extract. The names of components must year, month, day and hour}
}
\value{
A summary of your request and the link to download the data will be send via email.
In addition a summary of your request will be displayed on the console.
}
\description{
Function to request a subset of JRA55 minimum and maximum temperature at 2 meter above ground.
}
