% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtParseStationData_functions.R
\name{matchCDTStationsData}
\alias{matchCDTStationsData}
\title{Match CDT stations data.}
\usage{
matchCDTStationsData(...)
}
\arguments{
\item{...}{CDT stations data objects, possibly named.}
}
\value{
A list of CDT stations data objects, in same order as provided in the input arguments.
A named list if input arguments are named.
}
\description{
Filter CDT stations data to match the stations and dates.
}
