% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtUpdate_ONI_function.R
\name{oni.cpc.ncep.noaa}
\alias{oni.cpc.ncep.noaa}
\title{Get seasonal Oceanic Niño Index (ONI)}
\usage{
oni.cpc.ncep.noaa(startMonth, seasonLength)
}
\arguments{
\item{startMonth}{integer, start month of the season.}

\item{seasonLength}{integer, length of the season.}
}
\value{
A named list
}
\description{
Get seasonal Oceanic Niño Index from ONI data already downloaded in CDT.
}
\references{
\url{https://origin.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ONI_v5.php}
}
