% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtUpdate_ONI_function.R
\name{oni.iridl.ldeo}
\alias{oni.iridl.ldeo}
\title{Download seasonal Oceanic Niño Index (ONI)}
\usage{
oni.iridl.ldeo(startYear, seasonLength)
}
\arguments{
\item{startYear}{integer, start year to download.}

\item{seasonLength}{integer, length of the season.}
}
\value{
A named list
}
\description{
The seasonal Oceanic Niño Index is computed then downloaded from IRI Data Library.
}
\references{
\url{http://iridl.ldeo.columbia.edu/SOURCES/.NOAA/.NCDC/.ERSST/.version5}
}
