% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtDownloadReanalysis_era5_cds.R
\name{open_cds_profile}
\alias{open_cds_profile}
\title{Open CDS profile web page.}
\usage{
open_cds_profile()
}
\description{
Open Climate Data Store profile web page in the default browser.
}
