% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtPlot_MultipleData_function.R
\name{plotMultipleData}
\alias{plotMultipleData}
\title{Display maps from multiple data in one plot.}
\usage{
plotMultipleData(
  dataObj,
  shp = NULL,
  xlim = NULL,
  ylim = NULL,
  dataTitle = NULL,
  plotTitle = NULL,
  colorkeyTitle = "",
  presetCol = list(color = "tim.colors", reverse = FALSE),
  userCol = list(custom = FALSE, color = NULL),
  userLvl = list(custom = FALSE, levels = NULL, equidist = FALSE),
  ptsOpt = list(pch = 20, cex = 1),
  shpOpt = list(col = "black", lwd = 1)
)
}
\arguments{
\item{dataObj}{list of data. Each data has the form \code{list(x, y, z, type)}. \cr
Valid options for type: \code{"grid"} or \code{"points"}.
\itemize{
\item{\code{"grid"}: }{\code{x} vector, \code{y} vector, \code{z} matrix of \code{dim} equals \code{c(length(x), length(y))}}
\item{\code{"points"}: }{\code{x, y, z} vectors same length}
}}

\item{shp}{sp object of class \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}}

\item{xlim, ylim}{vector of length 2.}

\item{dataTitle}{vector same length as dataObj, title for each data}

\item{plotTitle}{character, title of the plot}

\item{colorkeyTitle}{character, title of the colorkey}

\item{presetCol}{list, preset colors for colorkey}

\item{userCol}{list, customized colors for colorkey}

\item{userLvl}{list, customized level for colorkey}

\item{ptsOpt}{list, options for points data}

\item{shpOpt}{list, options for the shp layer}
}
\description{
Display maps from multiple data in one plot.
}
