% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Humidity_functions.R
\name{potential_temperature}
\alias{potential_temperature}
\title{Potential temperature.}
\usage{
potential_temperature(tm, pr)
}
\arguments{
\item{tm}{numeric, a vector or matrix of the air temperature in degree Celsius.}

\item{pr}{numeric, a vector or matrix of the surface pressure in hPa.}
}
\value{
Returns a vector or matrix, the units are in degree Celsius.
}
\description{
Compute the potential temperature from the air temperature and surface pressure using the Poisson equation.
}
