% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_pressure_msl_conversion.R
\name{pressure_conversion_netcdf}
\alias{pressure_conversion_netcdf}
\title{Daily surface and mean sea level pressure conversion for a netCDF dataset.}
\usage{
pressure_conversion_netcdf(
  convert = "sfc2msl",
  start.date = "1991-01-01",
  end.date = "2020-12-31",
  pres.data = list(dir = "", format = "pres_\%s\%s\%s.nc", varid = "pres", ilon = 1, ilat
    = 2),
  temp.from = "minmax",
  tmax.data = list(dir = "", format = "tmax_\%s\%s\%s.nc", varid = "tmax", ilon = 1, ilat
    = 2),
  tmin.data = list(dir = "", format = "tmin_\%s\%s\%s.nc", varid = "tmin", ilon = 1, ilat
    = 2),
  tmean.data = list(dir = "", format = "tmean_\%s\%s\%s.nc", varid = "tmean", ilon = 1,
    ilat = 2),
  elev.data = list(file = "", varid = "z", ilon = 1, ilat = 2),
  output = list(dir = "", format = "prmsl_\%S.nc")
)
}
\arguments{
\item{convert}{character, the direction of conversion. Valid options: \code{"sfc2msl"} and \code{"msl2sfc"}.
\itemize{
\item{\code{"sfc2msl"}: }{conversion from surface to mean sea level pressure.}
\item{\code{"msl2sfc"}: }{conversion from mean sea level to surface pressure.}
}}

\item{start.date}{character, the start date of the data to be converted in the form \code{YYYY-MM-DD}.}

\item{end.date}{character, the end date of the data to be converted in the form \code{YYYY-MM-DD}.}

\item{pres.data}{named list, providing the pressure netCDF dataset. Units: hPa. 
Surface pressure or mean sea level pressure, depending on \code{convert}.
\itemize{
\item{\code{dir}: }{character, full path to the directory containing the netCDF files.}
\item{\code{format}: }{character, format of the netCDF file names}
\item{\code{varid}: }{character, name of the variable to read from the netCDF data}
\item{\code{ilon}: }{integer, order for the longitude dimension of the variable. 
Example: if the variable "pres" has the dimension order [Lat, Lon] then \code{ilon} must be 2}
\item{\code{ilat}: }{integer, order for the latitude dimension of the variable.}
}}

\item{temp.from}{character, source of temperature data. Valid options: \code{"minmax"} and \code{"mean"}.
\itemize{
\item{\code{"minmax"}: }{the mean temperature is computed from tme minimum and maximum temperature.}
\item{\code{"mean"}: }{the temperature data is from a mean temperature data.}
}}

\item{tmax.data}{named list, if \code{temp.from} is equal to \code{"minmax"}, 
providing the maximum temperature netCDF dataset. Units: degree Celsius.
See \code{pres.data} for the elements of the list.}

\item{tmin.data}{named list, if \code{temp.from} is equal to \code{"minmax"}, 
providing the minimum temperature netCDF dataset. Units: degree Celsius.
See \code{pres.data} for the elements of the list.}

\item{tmean.data}{named list, if \code{temp.from} is equal to \code{"mean"}, 
providing the mean temperature netCDF dataset. Units: degree Celsius.
See \code{pres.data} for the elements of the list.}

\item{elev.data}{named list, providing the Digital Elevation Model in netCDF format. Units: meters.
\itemize{
\item{\code{file}: }{character, full path to the netCDF file containing the elevation data.}
\item{\code{varid}: }{character, name of the variable to read from the netCDF data.}
\item{\code{ilon}: }{integer, order for the longitude dimension of the variable.}
\item{\code{ilat}: }{integer, order for the latitude dimension of the variable.}
}}

\item{output}{named list, the elements of the list are
\itemize{
\item{\code{dir}: }{character, full path to the directory to save the converted pressure}
\item{\code{format}: }{character, format of the output netCDF file names, the dates are represented by \code{\%S}}
}}
}
\description{
Function to convert surface and mean sea level pressure in both way.
}
