% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_pressure_msl_conversion.R
\name{pressure_conversion_station}
\alias{pressure_conversion_station}
\title{Daily surface and mean sea level pressure conversion for CDT station data format.}
\usage{
pressure_conversion_station(
  convert = "sfc2msl",
  pres.data = list(file = "", sep = ",", na.strings = "-99"),
  temp.data = "maxmin",
  tmax.data = list(file = "", sep = ",", na.strings = "-99"),
  tmin.data = list(file = "", sep = ",", na.strings = "-99"),
  tmean.data = list(file = "", sep = ",", na.strings = "-99"),
  elev.from = "inputPresData",
  elev.data = list(file = "", sep = ",", na.strings = "-99"),
  output = list(file = "", sep = ",", na.strings = "-99")
)
}
\arguments{
\item{convert}{character, the direction of conversion. Valid options: \code{"sfc2msl"} and \code{"msl2sfc"}.
\itemize{
\item{\code{"sfc2msl"}: }{conversion from surface to mean sea level pressure.}
\item{\code{"msl2sfc"}: }{conversion from mean sea level to surface pressure.}
}}

\item{pres.data}{named list, providing the pressure data in CDT station data format. Units: hPa.
\itemize{
\item{\code{file}: }{character, full path to the file containing the stations data}
\item{\code{sep}: }{character, column separator of the data}
\item{\code{na.strings}: }{character, missing values flag}
}}

\item{temp.data}{character, source of temperature data. Valid options: \code{"maxmin"} and \code{"mean"}.
\itemize{
\item{\code{"maxmin"}: }{the mean temperature is computed from tme minimum and maximum temperature.}
\item{\code{"mean"}: }{the temperature data is from a mean temperature data.}
}}

\item{tmax.data}{named list, if \code{temp.data} is equal to \code{"maxmin"}, 
providing the maximum temperature data in CDT station data format. Units: degree Celsius.
See \code{pres.data} for the elements of the list.}

\item{tmin.data}{named list, if \code{temp.data} is equal to \code{"maxmin"}, 
providing the minimum temperature data in CDT station data format. Units: degree Celsius.
See \code{pres.data} for the elements of the list.}

\item{tmean.data}{named list, if \code{temp.data} is equal to \code{"mean"}, 
providing the mean temperature data in CDT station data format. Units: degree Celsius.
See \code{pres.data} for the elements of the list.}

\item{elev.from}{character, source of the elevation data. Valid options: \code{"inputPresData"} and \code{"cdtCrdFile"}.
\itemize{
\item{\code{"inputPresData"}: }{the elevation data is from the pressure data.}
\item{\code{"cdtCrdFile"}: }{the elevation data is from a CDT coordinates file. The ID must be the same as the pressure data.}
}}

\item{elev.data}{named list, if \code{elev.from} is equal to \code{"cdtCrdFile"},
providing the CDT coordinate file containing the elevation data. Units: meters.
\itemize{
\item{\code{file}: }{character, full path to the CDT coordinate file}
\item{\code{sep}: }{character, column separator of the data}
\item{\code{na.strings}: }{character, missing values flag}
}}

\item{output}{named list, the elements of the list are
\itemize{
\item{\code{file}: }{character, full path to the file to save the converted pressure}
\item{\code{sep}: }{character, column separator of the data}
\item{\code{na.strings}: }{character, missing values flag}
}}
}
\description{
Function to convert surface and mean sea level pressure in both way.
}
