% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtParseStationData_functions.R
\name{readCDTStationData}
\alias{readCDTStationData}
\title{Read CDT station format file.}
\usage{
readCDTStationData(file, sep = ",", na.strings = "-99")
}
\arguments{
\item{file}{character, full path to the file containing the CDT data}

\item{sep}{character, the column's separator of the data}

\item{na.strings}{character, the missing values flag}
}
\value{
A list object
\itemize{
  \item{\strong{id}: }{Vector of the points/stations id}
  \item{\strong{lon}: }{Vector of the points/stations longitude}
  \item{\strong{lat}: }{Vector of the points/stations latitude}
  \item{\strong{dates}: }{Vector of the dates or times of the data}
  \item{\strong{data}: }{Matrix of the data, row indicates the dates and column the stations}
}
}
\description{
Read CDT station format file.
}
