% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtStatsProba_functions.R
\name{regression.Matrix}
\alias{regression.Matrix}
\title{Vectorization, regression}
\usage{
regression.Matrix(X, Y, min.len)
}
\arguments{
\item{X}{A numeric NxK matrix.}

\item{Y}{A numeric NxK matrix.}

\item{min.len}{An integer specify the minimum length of non missing data.}
}
\value{
Returns a numeric 10xK matrix.
}
\description{
Regression between two matrices performed by column \code{lm(Y~X)}.
}
\examples{
library(CDT)

set.seed(1)
X <- matrix(rnorm(200), nrow = 20, ncol = 10)
set.seed(2)
Y <- matrix(rnorm(200), nrow = 20, ncol = 10)
res <- regression.Matrix(X, Y, 10)

}
