% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Humidity_functions.R
\name{relative_humidity}
\alias{relative_humidity}
\title{Relative humidity.}
\usage{
relative_humidity(tm, td)
}
\arguments{
\item{tm}{numeric, a vector or matrix of the air temperature in degree Celsius.}

\item{td}{numeric, a vector or matrix of the dewpoint temperature in degree Celsius.}
}
\value{
Returns a vector or matrix, the units are in percentage.
}
\description{
Compute the relative humidity from the air temperature and dewpoint temperature.
}
