% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Humidity_functions.R
\name{saturation_vapour_pressure}
\alias{saturation_vapour_pressure}
\title{Saturation vapour pressure.}
\usage{
saturation_vapour_pressure(tmp)
}
\arguments{
\item{tmp}{numeric, a vector or matrix of the temperature data in degree Celsius.}
}
\value{
Returns a vector or matrix, the units are in hPa.
}
\description{
Compute the saturation vapour pressure.
}
