% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtDownloadReanalysis_era5_cds.R
\name{save_cds_personal_token}
\alias{save_cds_personal_token}
\title{Save CDS personal access token.}
\usage{
save_cds_personal_token(key)
}
\arguments{
\item{key}{CDS personal access token.}
}
\description{
Save CDS personal access token into a local file.
}
\examples{

\dontrun{
library(CDT)
save_cds_personal_token('10xasdxf2-ws12a-xdfs-zx3x-z1z2z3z4z5z')
}

}
