% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Humidity_functions.R
\name{specific_humidity}
\alias{specific_humidity}
\title{Specific humidity.}
\usage{
specific_humidity(td, pr)
}
\arguments{
\item{td}{numeric, a vector or matrix of the dewpoint temperature in degree Celsius.}

\item{pr}{numeric, a vector or matrix of the surface pressure in hPa.}
}
\value{
Returns a vector or matrix, the units are in kg/kg.
}
\description{
Compute the specific humidity from the dewpoint temperature and surface pressure.
}
