% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtGUI_StartUp.R
\name{startCDT}
\alias{startCDT}
\title{Start CDT GUI}
\usage{
startCDT(wd = NA, lang = NA)
}
\arguments{
\item{wd}{full path to the working directory.}

\item{lang}{the language to be used.}
}
\description{
Starting CDT GUI from R console.
}
\examples{

\dontrun{
library(CDT)

# Default working directory and English language
startCDT()

# Default working directory and Français language
startCDT(lang = "fr")

# Setting both working directory and language
startCDT("C:/Users/rijaf/Documents/AGHRYMET_ENACTS", "fr")
}

}
