% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtClimate_Humidity_functions.R
\name{temperature_from_theta}
\alias{temperature_from_theta}
\title{Air temperature.}
\usage{
temperature_from_theta(theta, pr)
}
\arguments{
\item{theta}{numeric, a vector or matrix of the potential temperature in degree Celsius.}

\item{pr}{numeric, a vector or matrix of the surface pressure in hPa.}
}
\value{
Returns a vector or matrix, the units are in degree Celsius.
}
\description{
Compute the air temperature from the potential temperature and surface pressure using the Poisson equation.
}
