% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtParseStationData_functions.R
\name{writeCDTStationData}
\alias{writeCDTStationData}
\title{Write CDT station format to a file.}
\usage{
writeCDTStationData(data, file, na.strings = "-99")
}
\arguments{
\item{data}{CDT stations data objects.}

\item{file}{character, full path to the file to save the CDT data}

\item{na.strings}{character, the missing values flag to write to the file}
}
\description{
Write CDT station format to a file.
}
