% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtBias_Options_functions.R
\name{biascoeff.options}
\alias{biascoeff.options}
\title{Options Controlling the bias coefficients parameters.}
\usage{
biascoeff.options(...)
}
\arguments{
\item{...}{using one or more arguments of the form name = value.
Existing values can be retrieved by supplying the names (as character strings) of the components as unnamed arguments.}
}
\description{
Functions to handle settings used by the bias coefficients computation.
}
