% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtPrecip_Merging_Cmd.R
\name{cdtMergingPrecipCMD}
\alias{cdtMergingPrecipCMD}
\title{Merging stations observation and satellite rainfall estimates data.}
\usage{
cdtMergingPrecipCMD(
  time.step = "dekadal",
  dates = list(from = "range", range = list(start = list(year = 2018, month = 1, dekad
    = 1), end = list(year = 2018, month = 12, dekad = 3)), file = "", dates = NULL),
  station.data = list(file = "", sep = ",", na.strings = "-99"),
  netcdf.data = list(dir = "", format = "rr_adj_\%s\%s\%s.nc", varid = "precip", ilon =
    1, ilat = 2),
  merge.method = list(method = "SBA", nrun = 3, pass = c(1, 0.75, 0.5)),
  interp.method = list(method = "idw", vargrd = FALSE, minstn = 10, nmin = 8, nmax =
    16, maxdist = 2.5, use.block = TRUE, vgm.model = c("Sph", "Exp", "Gau", "Pen")),
  auxvar = list(dem = FALSE, slope = FALSE, aspect = FALSE, lon = FALSE, lat = FALSE),
  dem.data = list(file = "", varid = "dem", ilon = 1, ilat = 2),
  grid = list(from = "data", netcdf = list(file = "", varid = "precip", ilon = 1, ilat
    = 2), bbox = list(minlon = 42, maxlon = 52, minlat = -26, maxlat = -11, reslon = 0.1,
    reslat = 0.1)),
  RnoR = list(use = FALSE, wet = 1, smooth = FALSE),
  blank = list(data = FALSE, shapefile = ""),
  output = list(dir = "", format = "rr_mrg_\%s\%s\%s.nc"),
  precision = list(from.data = TRUE, prec = "short"),
  GUI = FALSE
)
}
\arguments{
\item{time.step}{the time step of the data}

\item{dates}{a named list providing the dates to merge.
A component "from" with available option "range", "file" or "dates".
(to be completed)}

\item{station.data}{a named list (to be completed)}

\item{netcdf.data}{a named list (to be completed)}

\item{merge.method}{a named list (to be completed)}

\item{interp.method}{a named list (to be completed)}

\item{auxvar}{a named list (to be completed)}

\item{dem.data}{a named list (to be completed)}

\item{grid}{a named list (to be completed)}

\item{RnoR}{a named list (to be completed)}

\item{blank}{a named list (to be completed)}

\item{output}{a named list (to be completed)}

\item{precision}{a named list (to be completed)}

\item{GUI}{a logical indicating whether or not the output message should be displayed on CDT GUI.}
}
\description{
Function to merge stations observation and satellite rainfall estimates data.
}
