% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{cdtMergingTempCMD}
\alias{cdtMergingTempCMD}
\title{Merging Temperature}
\usage{
cdtMergingTempCMD(
  time.step = "dekadal",
  dates = list(from = "range", range = list(start = list(year = 2018, month = 1,
    dekad = 1), end = list(year = 2018, month = 12, dekad = 3)), file = "", dates = NULL),
  station.data = list(file = "", sep = ",", na.strings = "-99"),
  netcdf.data = list(dir = "", format = "tmax_adj_\%s\%s\%s.nc", varid = "temp", ilon =
    1, ilat = 2),
  merge.method = list(method = "SBA", nrun = 3, pass = c(1, 0.75, 0.5)),
  interp.method = list(method = "idw", vargrd = FALSE, minstn = 10, nmin = 8, nmax =
    16, maxdist = 3.5, use.block = TRUE, vgm.model = c("Sph", "Exp", "Gau", "Pen")),
  auxvar = list(dem = FALSE, slope = FALSE, aspect = FALSE, lon = FALSE, lat = FALSE),
  dem.data = list(file = "", varid = "dem", ilon = 1, ilat = 2),
  grid = list(from = "data", netcdf = list(file = "", varid = "precip", ilon = 1, ilat
    = 2), bbox = list(minlon = 42, maxlon = 52, minlat = -26, maxlat = -11, reslon = 0.1,
    reslat = 0.1)),
  blank = list(data = FALSE, shapefile = ""),
  output = list(dir = "", format = "tmax_mrg_\%s\%s\%s.nc"),
  GUI = FALSE
)
}
\arguments{
\item{time.step}{time step of the data}
\item{dates}{from: "range", "file", "dates"}
}
\description{
Function to merge temperature data.
}
