% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtTemp_Merging_Cmd.R
\name{cdtMergingTempCMD}
\alias{cdtMergingTempCMD}
\title{Merging stations observation and reanalysis data.}
\usage{
cdtMergingTempCMD(
  time.step = "dekadal",
  dates = list(from = "range", pars = list(start = "2018011", end = "2018123")),
  station.data = list(file = "", sep = ",", na.strings = "-99"),
  netcdf.data = list(dir = "", format = "tmax_adj_\%s\%s\%s.nc", varid = "temp", ilon =
    1, ilat = 2),
  merge.method = list(method = "SBA", nrun = 3, pass = c(1, 0.75, 0.5)),
  interp.method = list(method = "idw", nmin = 8, nmax = 16, maxdist = 3.5, use.block =
    TRUE, vargrd = FALSE, minstn = 10, vgm.model = c("Sph", "Exp", "Gau", "Pen")),
  auxvar = list(dem = FALSE, slope = FALSE, aspect = FALSE, lon = FALSE, lat = FALSE),
  dem.data = list(file = "", varid = "dem", ilon = 1, ilat = 2),
  grid = list(from = "data", pars = NULL),
  blank = list(data = FALSE, shapefile = ""),
  output = list(dir = "", format = "tmax_mrg_\%s\%s\%s.nc"),
  GUI = FALSE
)
}
\arguments{
\item{time.step}{the time step of the data. Available options: \code{"daily"}, \code{"pentad"}, \code{"dekadal"}, \code{"monthly"}.}

\item{dates}{a named list providing the dates to merge.
The list includes an element \code{from} with available options \code{"range"}, \code{"file"} or \code{"dates"}, and 
 an element \code{pars} which is a named list specifying the parameters related to \code{from}:
\itemize{
\item{\strong{"range"}: }{\code{pars} specifies the start and end dates to merge. \cr
Example: \code{pars = list(start = "2018011", end = "2018123")}}
\item{\strong{"file"}: }{\code{pars} specifies the full path to the file containing the dates to merge.
 Example: \code{pars = list(file = "/home/data/files/dates.txt")}\cr
The contents of the file are as follows:\cr
## cat /home/data/files/dates.txt \cr
2020011\cr
2020012\cr
2020013\cr
......
}
\item{\strong{"dates"}: }{\code{pars} specifies a vector containing the dates to merge. \cr
Example: \code{pars = list(dates = c("2020011", "2020012", 2020091, 2020113))}}
}}

\item{station.data}{a named list providing the station data to be used in CDT format.
\itemize{
\item{\code{file}: }{full path to the file containing the stations data}
\item{\code{sep}: }{column separator of the data}
\item{\code{na.strings}: }{missing values flag}
}}

\item{netcdf.data}{a named list providing the input netCDF dataset to be used.
\itemize{
\item{\code{dir}: }{full path to the directory containing the netCDF files.}
\item{\code{format}: }{format of the netCDF file names}
\item{\code{varid}: }{name of the variable to read from the netCDF data}
\item{\code{ilon}: }{order for the longitude dimension of the variable. 
Example: if the variable "temp" has the dimension order [Lat, Lon] then \code{ilon} must be 2}
\item{\code{ilat}: }{order for the latitude dimension of the variable.}
}}

\item{merge.method}{a named list indicating the merging method.
\itemize{
\item{\code{"method"}: }{the merging method. Valid options: \code{"CSc"}, \code{"BSc"}, \code{"SBA"} or \code{"RK"}.
 \itemize{
  \item{\strong{"CSc"}: }{Cressman Scheme}
  \item{\strong{"BSc"}: }{Barnes Scheme}
  \item{\strong{"SBA"}: }{Simple Bias Adjustment}
  \item{\strong{"RK"}: }{Regression Kriging}
 }
}
\item{\code{"nrun"}: }{number of the nested run to be performed}
\item{\code{"pass"}: }{vector giving the fraction of \code{nmin}, \code{nmax} and \code{maxdist} to be used for each pass.}
}}

\item{interp.method}{a named list indicating the interpolation method and parameters to be used for \code{"SBA"} and \code{"RK"}. 
\itemize{
\item{\code{method}: }{the interpolation method to be used.
Valid options: \code{"idw"}, \code{"shepard"}, \code{"sphere"} or \code{"okr"}.
\itemize{
 \item{\strong{"idw"}: }{Inverse distance weighted}
 \item{\strong{"shepard"}: }{Modified Shepard interpolation}
 \item{\strong{"sphere"}: }{Spheremap interpolation method}
 \item{\strong{"okr"}: }{Ordiranry kriging}
}
}
\item{\code{nmin}: }{minimum number of stations to be used to interpolate a grid point}
\item{\code{nmax}: }{maximum number of stations to be used to interpolate a grid point}
\item{\code{maxdist}: }{maximum radius of influence in decimal degree}
\item{\code{use.block}: }{logical, use block mean values to interpolate a grid point}
\item{\code{vargrd}: }{logical, use a variable radius of influence}
\item{\code{minstn}: }{minimum number of non missing values from station to perform the interpolation}
\item{\code{vgm.model}: }{vector of variogram model to be used if \code{method} is \code{"okr"}. Default is \code{c("Exp", "Gau", "Sph", "Pen")}}
}}

\item{auxvar}{a named list specifying the auxiliary variables to use when the merging method is \code{"RK"}.
\itemize{
\item{\code{dem}: }{logical, include elevation data as auxiliary variable}
\item{\code{slope}: }{logical, include slope as auxiliary variable}
\item{\code{aspect}: }{logical, include aspect as auxiliary variable}
\item{\code{lon}: }{logical, include longitude as auxiliary variable}
\item{\code{lat}: }{logical, include latitude as auxiliary variable}
}}

\item{dem.data}{a named list providing the Digital Elevation Model (in netCDF format) when using regression kriging method with elevation related data as auxiliary variable.
\itemize{
\item{\code{file}: }{full path to the netCDF file containing the elevation data.}
\item{\code{varid}: }{name of the variable to read from the netCDF data}
\item{\code{ilon}: }{order for the longitude dimension of the variable.}
\item{\code{ilat}: }{order for the latitude dimension of the variable.}
}}

\item{grid}{a named list providing the grid to use to interpolate the data.
The list includes an element \code{from} with available options \code{"data"}, \code{"ncdf"} or \code{"new"}, and 
 an element \code{pars} which is a named list specifying the parameters related to \code{from}:
\itemize{
\item{\strong{"data"}: }{the grid to interpolate will be taken from the input netCDF data, \code{pars} can be omitted} 
\item{\strong{"ncdf"}: }{the grid to interpolate will be taken from a provided netCDF file, 
\code{pars} specifies the full path to the netCDF file containing the grid to be extracted, 
the name of the variable to read, the order for the longitude and latitude dimension.\cr
Example: \code{pars = list(file = "/home/data/files/rfe_2020121.nc", varid = "precip", ilon = 1, ilat = 2)}
}
\item{\strong{"new"}: }{the grid to interpolate will be created from the information provided by user, 
\code{pars} specifies the minimum/maximum of the longitude and latitude of the domain and the resolution to be used. \cr
Example: \code{pars = list(minlon = 42, maxlon = 52, minlat = -26, maxlat = -11, reslon = 0.1, reslat = 0.1)}
}
}}

\item{blank}{a named list indicating if the data outside a provided shapefile will be removed.
\itemize{
\item{\code{data}: }{logical, blank the grid outside the provided shapefile}
\item{\code{shapefile}: }{full path to the shapefile to be used for blanking with the extension ".shp"}
}}

\item{output}{a named list indicating the directory to save the merged data and the format of the merged netCDF file name.}

\item{GUI}{a logical indicating whether or not the output message should be displayed on CDT GUI. If \code{TRUE}, CDT GUI must be open.}
}
\description{
Function to merge stations observation and reanalysis data.
}
