% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtExtractData_ncdf.R
\name{extract_netcdf_file}
\alias{extract_netcdf_file}
\title{Extract data from one netCDF file.}
\usage{
extract_netcdf_file(
  ncdf_data = list(file = "", varid = "z", ilon = 1, ilat = 2),
  crd_points = list(type = "cdtStnData", file = "", sep = ",", na.strings = "-99"),
  out_file = NULL
)
}
\arguments{
\item{ncdf_data}{named list, providing the input netCDF file to extract.
\itemize{
\item \code{file}: character, full path to the netCDF file.
\item \code{varid}: character, name of the variable to read from the netCDF data
\item \code{ilon}: integer, order for the longitude dimension of the variable. 
Example: if the variable "temp" has the dimension order [Lat, Lon] then \code{ilon} must be 2
\item \code{ilat}: integer, order for the latitude dimension of the variable.
}}

\item{crd_points}{named list, providing the coordinates over which the data will be extracted.
\itemize{
\item \code{type}: character, the type of file containing the coordinates of the points. Valid options: \code{"cdtStnData"}, \code{"cdtCrdFile"}.
 \itemize{
  \item \strong{"cdtStnData"}: the coordinates are from a CDT station data file
  \item \strong{"cdtCrdFile"}: the coordinates are from a CDT coordinates file
 }
\item \code{file}: character, full path to the file containing the coordinates
\item \code{sep}: character, column separator of the data
\item \code{na.strings}: character, missing values flag
}}

\item{out_file}{character, full path to the file in which the extracted data will be saved.
Default \code{NULL}, the extracted data will be returned and not saved to a file.}
}
\value{
If out_file is \code{NULL}, returns a data.frame with columns \code{id}, \code{lon}, \code{lat} and \code{value}.
}
\description{
Function to extract data over a set of points from one netCDF file.
}
