% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtInstall_wgrib.R
\name{install_wgrib}
\alias{install_wgrib}
\title{Install wgrib.}
\usage{
install_wgrib(cc = "gcc")
}
\arguments{
\item{cc}{the \code{C} compiler to be used or the full path to the \code{C} compiler.}
}
\description{
Function to download and compile wgrib into CDT local folder.
}
\examples{

\dontrun{
install_wgrib()
}

}
\references{
\url{https://www.cpc.ncep.noaa.gov/products/wesley/wgrib.html}
}
