% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtMerging_functions.R
\name{merging.options}
\alias{merging.options}
\title{Options Controlling the merging parameters.}
\usage{
merging.options(...)
}
\arguments{
\item{...}{using one or more arguments of the form name = value.
Existing values can be retrieved by supplying the names (as character strings) of the components as unnamed arguments.}
}
\description{
Functions to handle settings used by the merging.
}
\details{
Available options are
\itemize{ 
  \item{\code{saveGridBuffer}: }{logical, save the buffer of coarse grid used to interpolate the residuals. Default is \code{FALSE}}
  \item{\code{dirGridBuffer}: }{in case \code{saveGridBuffer} is \code{TRUE}, the full path to the directory to save the buffer data}
  \item{\code{saveRnoR}: }{logical, save the rain-no-rain mask. Default is \code{FALSE}}
  \item{\code{dirRnoR}: }{in case \code{saveRnoR} is \code{TRUE}, the full path to the directory to save the mask data}
  \item{\code{RnoRModel}: }{model to use to compute the rain-no-rain mask. Options are: \code{"logit"}, \code{"additive"}. Default is \code{"logit"}}
  \item{\code{RnoRCutOff}: }{integer, the method to be used to define the decision boundaries of the rain-no-rain mask. Options are: \code{1, 2, 3}
       \itemize{
          \item{\strong{option}: \code{1}}{
 %%
 \deqn{mask = \left\{
   \begin{array}{l l}
   0 & \quad \mbox{if } rnr < 0.5 \\
   1 & \quad \mbox{if } rnr \geq 0.5
   \end{array} \right.
 }{mask = if(rnr < 0) 0 else 1}
 %%
             }
          \item{\strong{option}: \code{2}}{
 %%
 \deqn{mask = \left\{
   \begin{array}{l l}
   0 & \quad \mbox{if } rnr < 0.1 \\
   rnr & \quad \mbox{if } rnr \geq 0.1
   \end{array} \right.
 }{mask = if(rnr < 0.1) 0 else rnr}
 %%
             }
          \item{\strong{option}: \code{3}}{
 %%
 \deqn{mask = \left\{
   \begin{array}{l l}
   0 & \quad \mbox{if } rnr < 0.25 \\
   rnr & \quad \mbox{if } 0.25 \leq rnr < 0.75 \\
   1 & \quad \mbox{if } rnr \geq 0.75
   \end{array} \right.
 }{mask = if(rnr < 0.25) 0 else if(rnr >= 0.25 & rnr < 0.75) rnr else 1}
 %%
               }
         }
    where \eqn{rnr} is the interpolated rain-no-rain values. Default is \code{3}.
    }
  \item{\code{RnoRaddCoarse}: }{logical, use the coarse grid to create the rain-no-rain mask. Default is \code{FALSE}}
  \item{\code{RnoRUseMerged}: }{logical, if \code{TRUE} the merged data is used to compute the rain-no-rain mask, otherwise the input gridded data is used. Default is \code{FALSE}}
  \item{\code{RnoRSmoothingPixels}: }{the number of pixels from the target pixel to be used to smooth the rain-no-rain mask. Default is 2.}
}
}
